% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creat_basis_intematlist.R
\name{create_basis_intematlist}
\alias{create_basis_intematlist}
\title{This function create the integrated bspline functions}
\usage{
create_basis_intematlist(timepoints, spline_basis)
}
\arguments{
\item{timepoints}{a list of observed time points}

\item{spline_basis}{the B-spline basis for the fitted FEC}
}
\description{
This function create the integrated bspline functions
}
\examples{
\dontrun{

}
}
