% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeFEC.R
\name{DeFEC}
\alias{DeFEC}
\title{This function computes the first K DeFEC}
\usage{
DeFEC(observed, timepoints, K = 3, spline_basis, basis_intematlist,
  thresh = 0.001, gammas = NULL, maxit = 500)
}
\arguments{
\item{observed}{a list of observed longitudinal data; each element in the list is a numeric vector for one subject's observed data}

\item{timepoints}{a list of observed time points}

\item{K}{the number of DeFEC}

\item{spline_basis}{the B-spline basis for the fitted FEC}

\item{basis_intematlist}{a list of integrated B-spline basis function}

\item{gammas}{a vector of length K, the smoothing parameter}
}
\description{
This function computes the first K DeFEC
}
\examples{
\dontrun{

}
}
