% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeFEC_plot.R
\name{DeFECplot}
\alias{DeFECplot}
\title{This function plots the estimated Derivatives}
\usage{
DeFECplot(DeFEC_pred, DeFEC_fd, sample_ids)
}
\arguments{
\item{DeFEC_pred}{results from function predict_DeFEC}

\item{DeFEC_fd}{results from function DeFEC}

\item{sample_ids}{sample id number}
}
\description{
This function plots the estimated Derivatives
}
\examples{
\dontrun{
}
}
