% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfda.R
\name{fdaggD}
\alias{fdaggD}
\title{This function plots the DeFECs}
\usage{
fdaggD(fdobj, ylab = "Value", xlab = "Time", addzero = FALSE,
  title = "", gridout = TRUE)
}
\arguments{
\item{fdobj}{DeFECs}
}
\description{
This function plots the DeFECs
}
\examples{
\dontrun{
}
}
