% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_residual.R
\name{pred_FOB}
\alias{pred_FOB}
\title{This function compute the residuals}
\usage{
pred_FOB(betalist, ylist, tlist, basis_intelist, spline_basis, nminus = 1,
  intercept = TRUE)
}
\arguments{
\item{betalist}{the coeffcients of the DeFEC}

\item{ylist}{a list of observed longitudinal data; each element in the list is a numeric vector for one subject's observed data}

\item{tlist}{a list of observed time points}

\item{basis_intelist}{a list of integrated B-spline basis function}

\item{spline_basis}{the B-spline basis for the fitted FEC}
}
\description{
This function compute the residuals
}
\examples{
\dontrun{

}
}
