% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_DeFEC.R
\name{predict_DeFEC}
\alias{predict_DeFEC}
\title{This function give the estimated derivative functions}
\usage{
predict_DeFEC(betalist, ylist, tlist, basis_intelist, spline_basis,
  nminus = 1)
}
\arguments{
\item{betalist}{the coeffcients of the DeFEC}

\item{ylist}{a list of observed longitudinal data; each element in the list is a numeric vector for one subject's observed data}

\item{tlist}{a list of observed time points}

\item{basis_intelist}{a list of integrated B-spline basis function}

\item{spline_basis}{the B-spline basis for the fitted FEC}
}
\description{
This function give the estimated derivative functions
}
\examples{
\dontrun{

}
}
