% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_gamma_FirstDeFEC.R
\name{select_gamma}
\alias{select_gamma}
\title{This function computes the cv errors for different smoothing parameter}
\usage{
select_gamma(gamma_pool, folds, obs, timep, basis_intelist, spline_basis,
  thres = 1e-04, betalist = NULL, evalc = 5, maxit = 1000)
}
\arguments{
\item{gamma_pool}{a vector of canadiate smoothing parameter values}

\item{folds}{cv folds}

\item{obs}{observed response, a list}

\item{basis_intelist}{a list of integrated B-spline basis function}

\item{spline_basis}{the B-spline basis for the fitted FEC}

\item{thres}{threshold}

\item{betalist}{previous DeFEC's coefs}

\item{evalc}{evaluate times}

\item{maxit}{maxit}

\item{timp}{timepoints corresponding to the observed response, a list}
}
\description{
This function computes the cv errors for different smoothing parameter
}
\examples{
\dontrun{
}
}
