% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_pc_step.R
\name{single_FPC}
\alias{single_FPC}
\title{This function computes the first DeFEC}
\usage{
single_FPC(betalist = NULL, obs, timep, basis_intelist, spline_basis,
  threshold = 0.001, minit = 20, gamma = 0, maxeval = 1000)
}
\arguments{
\item{betalist}{a list of previous DeFEC's coefficients, the resulting one should be orthogonal to those previous computed ones.}

\item{obs}{a list of observed longitudinal data; each element in the list is a numeric vector for one subject's observed data}

\item{timep}{a list of observed time points}

\item{basis_intelist}{a list of integrated B-spline basis function}

\item{spline_basis}{the B-spline basis for the fitted FEC}

\item{gamma}{a positive number, the smoothing parameter}
}
\description{
This function computes the first DeFEC
}
\examples{
\dontrun{

}
}
